+++ resolwe_bio/kb/management/commands/insert_features.py
@@ -13,7 +13,6 @@ from django.core.exceptions import ValidationError
 from django.core.management.base import BaseCommand
 from django.db import connection
 
-from resolwe.elastic.builder import index_builder
 from resolwe.utils import BraceMessage as __
 
 from resolwe_bio.kb.models import Feature
@@ -188,8 +187,6 @@ class Command(BaseCommand):
             count_total += len(features)
             count_inserted += result[1]
             count_updated += result[2]
-
-        index_builder.build(queryset=Feature.objects.filter(id__in=to_index))
 
         logger.info(
             "Total features: %d. Inserted %d, updated %d, unchanged %d."