+++ tests/urls.py
@@ -3,17 +3,18 @@ from django.urls import include, path
 from rest_framework import routers
 
 from resolwe.api_urls import api_router as resolwe_router
-from resolwe.elastic import routers as search_routers
 from resolwe.flow.views import EntityViewSet
 from resolwe_bio.kb.views import (FeatureSearchViewSet, FeatureAutocompleteViewSet, FeatureViewSet,
                                   MappingViewSet, MappingSearchViewSet)
+from .routers import SearchRouter
+
 
 api_router = routers.DefaultRouter(trailing_slash=False)  # pylint: disable=invalid-name
 api_router.register(r'sample', EntityViewSet)
 api_router.register(r'kb/feature/admin', FeatureViewSet)
 api_router.register(r'kb/mapping/admin', MappingViewSet)
 
-search_router = search_routers.SearchRouter(trailing_slash=False)  # pylint: disable=invalid-name
+search_router = SearchRouter(trailing_slash=False)  # pylint: disable=invalid-name
 search_router.register(r'kb/feature/search', FeatureSearchViewSet, 'kb_feature_search')
 search_router.register(r'kb/feature/autocomplete', FeatureAutocompleteViewSet, 'kb_feature_autocomplete')
 search_router.register(r'kb/mapping/search', MappingSearchViewSet, 'kb_mapping_search')