+++ docs/CHANGELOG.rst
@@ -17,6 +17,8 @@ Changed
 -------
 - Bump Samtools to version 1.10 in ``resolwebio/common:1.6.0`` Docker
   image
+- Migrate search for Knowledge Base enpoints from Elasticsearch to
+  PostgreSQL
 
 
 ===================